refactor(engine): split engine.js into per-phase files in engine/ folder
